refactor: 당첨 결과 생성 역할 분리
