+++ src/main/kotlin/lottery/domain/LotteryMachine.kt
@@ -1,7 +1,7 @@
 package lottery.domain
 
 object LotteryMachine {
-    fun buy(tickets: List<Ticket>): List<Lottery> {
+    fun purchase(tickets: List<Ticket>): List<Lottery> {
         return List(tickets.size) { Lottery.create() }
     }
 }
