+++ src/main/kotlin/lottery/domain/WinningLottery.kt
@@ -1,10 +1,6 @@
 package lottery.domain
 
-class WinningLottery(val lottery: Lottery) {
-    fun draw(lotteries: List<Lottery>): DrawResult {
-        return DrawResult.from(winningLottery = this, lotteries = lotteries)
-    }
-
+data class WinningLottery(val lottery: Lottery) {
     companion object {
         fun create(numbers: List<Int>): WinningLottery {
             return WinningLottery(Lottery(LotteryNumbers(numbers.toSet())))
