+++ src/test/kotlin/lottery/domain/LotteryMachineTest.kt
@@ -7,7 +7,7 @@ import io.kotest.matchers.shouldBe
 class LotteryMachineTest : StringSpec({
     "티켓 수 만큼 로또를 생성한다" {
         listOf(1, 2, 3, 4, 5).forAll { count ->
-            LotteryMachine.buy(List(count) { Ticket() }).size shouldBe count
+            LotteryMachine.purchase(List(count) { Ticket() }).size shouldBe count
         }
     }
 })
